Increase maintainability by shifthing all get_ functions to get_utils.
